requirements.txt
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-JWT-Extended==3.21.0
waitress==1.4.4
nose2==0.9.1
passlib==1.7.1
psycopg2-binary==2.8.4
marshmallow-sqlalchemy==0.23.1
flask-marshmallow==0.13.0
flask-migrate==2.5.3